Fixed issues identified with static analysis

+++ common/alignment.c
@@ -203,9 +203,9 @@ void alignment_print_spacer(const char* alignment_a, const char* alignment_b,
 
 // Backtrack through scoring matrices
 void alignment_reverse_move(enum Matrix *curr_matrix, score_t* curr_score,
-                            unsigned int *score_x, unsigned int *score_y,
+                            size_t *score_x, size_t *score_y,
                             unsigned long *arr_index,
-                            unsigned int score_width, unsigned int score_height,
+                            size_t score_width, size_t score_height,
                             const score_t *match_score,
                             const score_t *gap_a_score,
                             const score_t *gap_b_score,
@@ -214,8 +214,8 @@ void alignment_reverse_move(enum Matrix *curr_matrix, score_t* curr_score,
 {
   int prev_match_penalty, prev_gap_a_penalty, prev_gap_b_penalty;
 
-  unsigned int seq_x = (*score_x)-1;
-  unsigned int seq_y = (*score_y)-1;
+  size_t seq_x = (*score_x)-1;
+  size_t seq_y = (*score_y)-1;
 
   int match_penalty = scoring_lookup(scoring, seq_a[seq_x], seq_b[seq_y]);
